change folder name tests
